Properly search through paginated people results
Previously, the Planning Center API was only returning an arbitrary 25
people (the default pagination results-per-page count). Because the
filtering was being done *after* the API returned, the workflow could
not access the vast majority of people for that Planning Center account.

To solve this, switch to performing the filtering as part of the initial
GET request (which I discovered I could do via Planning Center's API;
wish I knew that sooner).